+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Add Sandhole version to admin interface.
+
 ## 0.2.0 (2024-12-20)
 
 ### Added